refactor: useCommonMenuBarDataを移動
